+++ ios/Tests/MeasureSDKTests/Mocks/MockSystemFileManager.swift
@@ -36,4 +36,22 @@ final class MockSystemFileManager: SystemFileManager {
     func getDynamicConfigPath() -> String? {
         return dynamicConfigPath
     }
+
+    func retrieveFile(atPath path: String) -> Data? {
+        return savedFiles[path]
+    }
+
+    func deleteFile(atPath path: String) {
+        savedFiles.removeValue(forKey: path)
+    }
+
+    func cleanupOrphanedAttachmentFiles(validPaths: Set<String>) {
+        let allPaths = Array(savedFiles.keys)
+
+        for path in allPaths {
+            if !validPaths.contains(path) {
+                savedFiles.removeValue(forKey: path)
+            }
+        }
+    }
 }