+++ react-native/example/rnExample/App.tsx
@@ -13,9 +13,7 @@ import {
 import {Colors} from 'react-native/Libraries/NewAppScreen';
 import {
   Measure,
-  ClientInfo,
   MeasureConfig,
-  ScreenshotMaskLevel,
 } from '@measuresh/react-native';
 
 type ActionItem = {
@@ -36,31 +34,12 @@ const App = (): React.JSX.Element => {
   const [showScreenshot, setShowScreenshot] = useState(false);
 
   const initializeMeasure = async () => {
-    const clientInfo = new ClientInfo(
-      'msrsh_38514d61493cf70ce99a11abcb461e9e6d823e2068c7124a0902b745598f7ffb_65ea2c1c',
-      'msrsh_38514d61493cf70ce99a11abcb461e9e6d823e2068c7124a0902b745598f7ffb_65ea2c1c',
-      'http://localhost:8080',
-    );
-
     const measureConfig = new MeasureConfig({
       enableLogging: true,
-      samplingRateForErrorFreeSessions: 1.0,
-      coldLaunchSamplingRate: 1.0,
-      warmLaunchSamplingRate: 1.0,
-      hotLaunchSamplingRate: 1.0,
-      userJourneysSamplingRate: 1.0,
-      traceSamplingRate: 1.0,
-      trackHttpHeaders: true,
-      trackHttpBody: true,
-      httpHeadersBlocklist: [],
-      httpUrlBlocklist: [],
-      httpUrlAllowlist: [],
       autoStart: true,
-      screenshotMaskLevel: ScreenshotMaskLevel.allText,
-      maxDiskUsageInMb: 50,
     });
 
-    await Measure.init(clientInfo, measureConfig);
+    await Measure.init(measureConfig);
 
     Measure.onShake(() => {
       console.log('Shake detected — launching bug report flow!');
@@ -74,18 +53,10 @@ const App = (): React.JSX.Element => {
 
   const startMeasure = () => {
     Measure.start()
-      .then(() => console.log('Measure SDK started successfully'))
-      .catch((error: any) =>
-        console.error('Failed to start Measure SDK:', error),
-      );
   };
 
   const stopMeasure = () => {
     Measure.stop()
-      .then(() => console.log('Measure SDK stopped successfully'))
-      .catch((error: any) =>
-        console.error('Failed to stop Measure SDK:', error),
-      );
   };
 
   const trackCustomEvent = () => {