+++ react-native/src/config/screenshotMaskLevel.ts
@@ -6,21 +6,21 @@ export enum ScreenshotMaskLevel {
    * The strictest level of masking which masks all text,
    * input fields, images, and videos.
    */
-  allTextAndMedia = "AllTextAndMedia",
+  allTextAndMedia = "all_text_and_media",
 
   /**
    * Masks all text and input fields, including clickable elements.
    */
-  allText = "AllText",
+  allText = "all_text",
 
   /**
    * Masks all text and input fields, excluding clickable elements.
    */
-  allTextExceptClickable = "AllTextExceptClickable",
+  allTextExceptClickable = "all_text_except_clickable",
 
   /**
    * The most lenient masking level which only masks sensitive input fields
    * like passwords, email fields, and phone number fields.
    */
-  sensitiveFieldsOnly = "SensitiveFieldsOnly"
+  sensitiveFieldsOnly = "sensitive_fields_only"
 }