+++ CHANGELOG.md
@@ -87,6 +87,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### :books: Documentation
 
+- Clarify http headers collection (#3133) by @abhaysood in #3133
 - Update outdated sdk docs (#3004) by @abhaysood in #3004
 - Fix placeholder key for API URL (#2896) by @Shabinder in #2896